Take group out of transfer structure
We'll make group a first-class component of all transfer paths
instead. This makes it clear that transfers always belong to a
group and should never be accessed without one. It also implies
you know the group of any transfer you are accessing, so there
is no need to include it in the serialization.