Ensure we update updated_at when marking progress
Sequel::Model#update is smart enough not to update anything if nothing has changed,
but this becomes a liability here since in the course of, e.g., an index creation,
the processed_bytes can stay stagnant for a relatively long time. This makes sure
we always update updated_at from Transfer#mark_progress.

Note that actually communicating that something is still happening in a situation
like this is another, separate problem.